jobs/requirements.txt
playwright
bs4
dotenv
kafka-python
pyspark
logging
cassandra-driver